requirements.txt
torch
torchvision
numpy
pandas
Pillow
open_clip_torch
tensorboard
